+++ src/Scheme.php
@@ -24,15 +24,18 @@ use League\Url\Utilities;
  */
 class Scheme extends Component implements Interfaces\Scheme
 {
-    use Utilities\StandardPort;
+    use Utilities\RegisteredSchemes;
 
     /**
      * {@inheritdoc}
      */
     protected function validate($data)
     {
-        if (! static::isSupported($data)) {
-            throw new InvalidArgumentException('Unsupported scheme');
+        if (! static::isRegistered($data)) {
+            throw new InvalidArgumentException(sprintf(
+                "the submitted scheme '%s' is no registered you should use `Scheme::register` first",
+                $data
+            ));
         }
 
         return strtolower($data);
@@ -41,20 +44,81 @@ class Scheme extends Component implements Interfaces\Scheme
     /**
      * Tell wether the submitted scheme is implemented in the package
      *
-     * @param  string  $scheme
+     * @param string $scheme
      *
-     * @throws InvalidArgumentException If the submitted data is not implemented
+     * @throws InvalidArgumentException If the submitted scheme is invalid
      *
      * @return boolean
      */
-    public static function isSupported($scheme)
+    public static function isRegistered($scheme)
     {
-        if (! preg_match('/^[a-z][-a-z0-9+.]+$/i', $scheme)) {
-            throw new InvalidArgumentException('The submitted data is invalid');
+        return isset(static::$registeredSchemes[static::formatScheme($scheme)]);
+    }
+
+    /**
+     * Validate Scheme syntax according to RFC3986
+     *
+     * @param string $scheme
+     *
+     * @throws InvalidArgumentException If the submitted scheme is invalid
+     *
+     * @return string
+     */
+    protected static function formatScheme($scheme)
+    {
+        if (! filter_var($scheme, FILTER_VALIDATE_REGEXP, ['options' => ['regexp' => '/^[a-z][-a-z0-9+.]+$/i']])) {
+            throw new InvalidArgumentException(sprintf("Invalid Submitted scheme: '%s'", $scheme));
         }
-        $res = explode('+', $scheme);
-        $real_scheme = strtolower(array_pop($res));
-        return empty($real_scheme) || isset(static::$standardPorts[$real_scheme]);
+
+        return strtolower($scheme);
+    }
+
+    /**
+     * Register a new Scheme or add standard Port to an
+     * already registered Scheme
+     *
+     * @param  string   $scheme
+     * @param  int|null $port
+     *
+     * @throws InvalidArgumentException If the submitted port is invalid
+     */
+    public static function register($scheme, $port = null)
+    {
+        $scheme = static::formatScheme($scheme);
+        static::assertValidScheme($scheme);
+        if (! isset(static::$registeredSchemes[$scheme])) {
+            static::$registeredSchemes[$scheme] = [];
+        }
+        if (empty($port)) {
+            return;
+        }
+        if (! filter_var($port, FILTER_VALIDATE_INT, ['options' => ['min_range' => 1, 'max_range' => 65535]])) {
+            throw new InvalidArgumentException('The submitted port is invalid');
+        };
+        static::$registeredSchemes[$scheme][] = $port;
+        sort(static::$registeredSchemes[$scheme]);
+    }
+
+    /**
+     * Restrict action on default Schemes
+     *
+     * @param string $scheme
+     *
+     * @throws InvalidArgumentException If the scheme is part
+     *                                     of the defaults list of registered scheme
+     */
+    protected static function assertValidScheme($scheme)
+    {
+        if (in_array($scheme, ['http', 'https', 'ws', 'wss', 'ftp', 'ftps', 'file'])) {
+            throw new InvalidArgumentException(sprintf("The submitted scheme '%s' is already defined", $scheme));
+        }
+    }
+
+    public static function unRegister($scheme)
+    {
+        $scheme = static::formatScheme($scheme);
+        static::assertValidScheme($scheme);
+        unset(static::$registeredSchemes[$scheme]);
     }
 
     /**
@@ -62,14 +126,15 @@ class Scheme extends Component implements Interfaces\Scheme
      */
     public function getStandardPorts()
     {
-        $res = [];
-        if (! $this->isEmpty()) {
-            $res = explode('+', $this->data);
+        $res  = [];
+        if (isset(static::$registeredSchemes[$this->data])) {
+            $res = static::$registeredSchemes[$this->data];
         }
+        sort($res);
 
         return array_map(function ($value) {
             return new Port($value);
-        }, $this->getAssociatedData(static::$standardPorts, array_pop($res)));
+        }, $res);
     }
 
     /**